+++ include/optimize_cmp_kmers/read_in_kmer.hpp
@@ -16,7 +16,7 @@ constexpr double pseudo_count_amount = 1.0;
 namespace container{
  
 struct RI_Kmer{
-    Eigen::Array4d next_char_prob;
+    Eigen::Array4f next_char_prob;
     //Could be 3 and infer the fourth probability,
     //Using 4 since it may be useful in SIMD instructions
     // size_t length = 0; 
@@ -27,11 +27,11 @@ struct RI_Kmer{
 
     RI_Kmer(const vlmc::VLMCKmer &old_kmer){
         // this->length = old_kmer.length;
-        Eigen::Array4d tmp = {old_kmer.next_symbol_counts[0], 
+        Eigen::Array4f tmp = {old_kmer.next_symbol_counts[0], 
                               old_kmer.next_symbol_counts[1],
                               old_kmer.next_symbol_counts[2],
                               old_kmer.next_symbol_counts[3]};
-        double child_count = tmp.sum() + 4;
+        float child_count = tmp.sum() + 4;
         this->next_char_prob = (tmp + pseudo_count_amount) / child_count;
         this->integer_rep = get_index_rep(old_kmer);
         this->is_null = false;
@@ -41,7 +41,7 @@ struct RI_Kmer{
     RI_Kmer(const std::array<double, 4>& next_counts){ // , const int len){
         // this->length = len;
         int pseudo_count_amount = 1;
-        double child_count = std::accumulate(next_counts.begin(), next_counts.end(), pseudo_count_amount * 4);
+        float child_count = std::accumulate(next_counts.begin(), next_counts.end(), pseudo_count_amount * 4);
         this->next_char_prob = {(double(next_counts[0]) + pseudo_count_amount) / child_count,
                (double(next_counts[1]) + pseudo_count_amount) / child_count,
                (double(next_counts[2]) + pseudo_count_amount) / child_count,
